requirements.txt
about-time==4.2.1
alive-progress==3.1.5
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.4.0
autopep8==2.1.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.10.5
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.7
comm==0.2.0
contourpy==1.3.3
cryptography==37.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
grapheme==0.6.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.1
lxml==4.9.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.6
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.1.0
ply==3.11
prettytable==3.10.0
prometheus_client==0.23.1
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.20
Pygments==2.17.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==4.0.0
pytz==2024.1
PyYAML==6.0
pyzmq==25.1.2
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.21
tinycss2==1.4.0
torch==2.9.0
torchvision==0.24.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
triton==3.5.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.12
wcwidth==0.2.12
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.14